Proxy: change instantiation of the object mapper
Configuring the object mapper must be done before instantiating it, with
a few exception. Its current configuration allowed it to use reflection
on types to find getters and add their returned values in the json,
which we want to prevent as all the data should be found in fields only
and allow us to add utility methods in proxy classes.