Proxy: Change Role's type field to TypeExprs
Getters and Setters for array types need to have a role in the proxy.
However, there is no declaration for the array type, and we need to know
for which underlying element type the function is used for. TypeExprs
are able to better specify that information.

Only Arrays and Typenames should be found in the Role's type field (it
does not make much sense for a function to be the method of a reference
type...), so this commit also adds a check in the validator.