Proxy: Add the project name in the proxy
For binded libraries to not clash, we need to keep the name of the
binded library in the proxy.

For example, exception types will need an enumeration and a function to
identify which exception is being thrown in order to ensure that the
correct type is kept between binding calls. Since multiple libraries
will try to generate similar functions - with clashing names -, we need
to namespace them to their corresponding libraries.