Ada2Proxy: Add copy function
The action of being able to copy classes was not yet supported. This
meant that copying values in target languages could lead to unwanted
behaviors: in the current printer available (C++), copying an object
would result in the copy of the data pointer, ultimately leading to
a double free error. The C++ Rule of 3 is now supported for binded
types.

This also will be used for supporting controlled types in the future,
notably the `Adjust` primitive.