Ada2Proxy: add with-clause for renamed package dependencies and runtime packages
With-clauses are not transitive when the with-ed package is a simple
renaming of an other. Since package renaming does not recreate entities,
using the fully qualified name of an entity does not ensure that it can
be found in the case of a renamed package dependency. We thus need to
add a with-clause for the renamed package.