Ada2Proxy: Make unique names only in the vtable record
The vtable entries in the proxy must have the real names of the methods
they map to. Otherwise, it makes it impossible to know which functions
they should dispatch to.

The names of each in the record vtable do not matter, only the order of
the field does, which is reflected in the order of the vtable entries in
the proxy. So we only need to ensure that names of fields are unique in
the record.