Ada2Proxy: Rework array bindings gen
The way arrays were generated proved to be problematic when their
component type is a controlled object type.

All access types to controlled types, or types containing controlled
types hold references to all heap allocated values, which are freed when
the type goes out of scope. When returning an array, an access type was
created to be able to return a dynamically allocated copy. However,
this meant that right before returning the access value, the memory
pointed was finalized.

This commit adds the presence anywhere in the AdaProxy declaration
lists, instead of only storing their component type as a Set. This
allows to create access types in the global scope which do not free
arrays when they are returned. Array functions (getter, setters...) are
created by visiting the entire proxy, and getting a list of all
array components used in the proxy.